Added conversion to int for second to millisecond conversion

+++ SpatialGDK/Source/SpatialGDK/Private/Interop/Connection/SpatialConnectionManager.cpp
@@ -100,10 +100,10 @@ struct ConfigureConnection
 		const USpatialGDKSettings* Settings = GetDefault<USpatialGDKSettings>();
 
 #if WITH_EDITOR
-		Worker_HeartbeatParameters WithEditorHeartbeatParams = { Settings->HeartbeatIntervalSeconds * SECONDS_TO_MILLIS,
-														 Settings->HeartbeatTimeoutWithEditorSeconds * SECONDS_TO_MILLIS };
+		Worker_HeartbeatParameters WithEditorHeartbeatParams = { (uint64_t) Settings->HeartbeatIntervalSeconds * SECONDS_TO_MILLIS,
+																 (uint64_t) Settings->HeartbeatTimeoutWithEditorSeconds * SECONDS_TO_MILLIS };
 		Params.network.tcp.downstream_heartbeat = &WithEditorHeartbeatParams;
-		Params.network.tcp.upstream_heartbeat = &WithEditorHeartbeatParams; 
+		Params.network.tcp.upstream_heartbeat = &WithEditorHeartbeatParams;
 		Params.network.kcp.downstream_heartbeat = &WithEditorHeartbeatParams;
 		Params.network.kcp.upstream_heartbeat = &WithEditorHeartbeatParams;
 #else
